rename functions folder
